Add support for loading the real Variables for an Expression
When we parse an Expression, we search the table chain for the referenced
variables (and crash if it's not found...) and then prepare the Expression
to be displayed as part of the parse tree.

Later on, this will be transformed into creating a quad that represents the
operation and pushing it to the function's statements list.